requirements.txt
flask
flask_restx
flask_cors
sqlalchemy
nltk
sastrawi
numpy
python-dotenv